ImageParser/requirements.txt
Flask==2.1.0
numpy==1.24.0
torch==1.13.1
torchvision==0.14.1
python-doctr==0.6.0
timm
requests